Added more functionalities to events pages like changing, removing and subscribing to event

+++ Ignite/Ignite.Services/Events/EventsService.cs
@@ -1,4 +1,5 @@
 ﻿using Ignite.Data;
+using Ignite.Infrastructure.CustomAttributes;
 using Ignite.Models;
 using Ignite.Models.InputModels.Events;
 using Ignite.Models.ViewModels.Events;
@@ -6,6 +7,7 @@ using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -47,6 +49,22 @@ namespace Ignite.Services.Events
             db.SaveChanges();
         }
 
+        public void AddUserToEvent(string userId, string eventId)
+        {
+            if (!db.Users.Any(u => u.Id == userId) || !CheckEventExists(eventId))
+            {
+                throw new ArgumentException("Invalid data.");
+            }
+
+            db.UsersEvents.Add(new UserEvent
+            {
+                EventId = eventId,
+                UserId = userId
+            });
+
+            db.SaveChanges();
+        }
+
         public bool CheckEventExists(string eventId)
         {
             return db.Events.Any(x => x.Guid == eventId);
@@ -58,7 +76,7 @@ namespace Ignite.Services.Events
                 .Where(e => !e.IsDeleted)
                 .Include(e => e.UsersEvents)
                 .Select(e => new ShowEventsViewModel
-                { 
+                {
                     Guid = e.Guid,
                     Name = e.Name,
                     Address = e.Address,
@@ -71,8 +89,69 @@ namespace Ignite.Services.Events
 
         public void RemoveEvent(string eventId)
         {
+            if(!CheckEventExists(eventId))
+                throw new ArgumentException("Invalid data.");
+
             db.Events.Remove(db.Events.Find(eventId));
             db.SaveChanges();
+        }
+
+        public void RemoveUserFromEvent(string userId, string eventId)
+        {
+            if (!db.Users.Any(u => u.Id == userId) || !CheckEventExists(eventId))
+            {
+                throw new ArgumentException("Invalid data.");
+            }
+
+            db.UsersEvents.Remove(db.UsersEvents.First(ue => ue.EventId == eventId && ue.UserId == userId));
+
+            db.SaveChanges();
+        }
+
+        public void ChangeEvent(Models.InputModels.Events.ChangeEventInputModel model)
+        {
+            if (!CheckEventExists(model.Guid))
+                throw new ArgumentException("Invalid data.");
+
+            if (string.IsNullOrWhiteSpace(model.Name))
+            {
+                throw new ArgumentException("Name must be provied!");
+            }
+            else if (string.IsNullOrWhiteSpace(model.Address))
+            {
+                throw new ArgumentException("Address must be provied!");
+            }
+            else if (model.StartingDateTime < DateTime.Now)
+            {
+                throw new ArgumentException("Invalid Date & Time!");
+            }
+
+            var ev = GetEventByGUID(model.Guid);
+
+            var evProperties = ev.GetType().GetProperties();
+            var modelProperties = model.GetType().GetProperties();
+
+            foreach (var propM in modelProperties)
+            {
+                foreach (var propE in evProperties)
+                {
+                    if (propE.PropertyType == propM.PropertyType &&
+                        propE.Name == propM.Name)
+                    {
+                        propE.SetValue(ev, propM.GetValue(model));
+                    }
+                }
+            }
+
+            db.SaveChanges();
+        }
+
+        public Event GetEventByGUID(string eventId)
+        {
+            if(!CheckEventExists(eventId))
+                throw new ArgumentException("Invalid data.");
+
+            return db.Events.First(e => e.Guid == eventId);
         }
     }
 }
